Staging

+++ app/Http/Models/Deal.php
@@ -54,8 +54,8 @@ class Deal extends Model
 	protected $casts = [
 		'id_product' => 'int',
 		'deals_voucher_duration' => 'int',
-		'deals_voucher_price_point' => 'int',
-		'deals_voucher_price_cash' => 'int',
+		'deals_voucher_price_point' => 'double',
+		'deals_voucher_price_cash' => 'double',
 		'deals_total_voucher' => 'int',
 		'total_voucher_subscription' => 'int',
 		'deals_total_claimed' => 'int',
